requirements.txt
click>=7.0,<8
flask>=0.12.5,<0.13
pyyaml>=5.4.1,<6
voluptuous>=0.12.1,<0.13
